+++ src/services/lunchService.js
@@ -32,14 +32,18 @@ const lunchDelete = async (id) => {
 		return query.affectedRows; // affectedRows가 1인지 아닌지 확인하기 위함 
 };
 
-const lunchSlack = async () => {
+const slackSend = async (startDate, endDate) => {
 	const connection = await pool.getConnection(async conn => conn);
-	const [rows] = await connection.query("SELECT name, food FROM lunch WHERE updatedAt BETWEEN ")
-}
+	const [rows] = await connection.query("SELECT name, food FROM lunch WHERE updatedAt BETWEEN ? AND ?", [startDate, endDate])
+
+	return rows
+};
+
 module.exports = {
 	lunchList,
 	lunchView,
 	lunchInput,
 	lunchUpdate,
-	lunchDelete
+	lunchDelete,
+	slackSend
 };
